fixed int32 overflow on gps lat/lon

+++ src/drivers/uavcannode/UavcanNode.cpp
@@ -391,8 +391,8 @@ void UavcanNode::Run()
 
 			fix2.gnss_time_standard = fix2.GNSS_TIME_STANDARD_UTC;
 			fix2.gnss_timestamp.usec = gps.time_utc_usec;
-			fix2.latitude_deg_1e8 = gps.lat * 10;
-			fix2.longitude_deg_1e8 = gps.lon * 10;
+			fix2.latitude_deg_1e8 = (int64_t)gps.lat * 10;
+			fix2.longitude_deg_1e8 = (int64_t)gps.lon * 10;
 			fix2.height_msl_mm = gps.alt;
 			fix2.height_ellipsoid_mm = gps.alt_ellipsoid;
 			fix2.status = gps.fix_type;
